feature(nemesis): node is banned if it was removed from cluster
New nemesis for raft feature consistent-topology-changes:
ban removed node if it try back to cluster
The new nemesis send signal to scylla process and set it PAUSED
After the node with paused scylla marked as DN, the node is removed
After that signal SIGCONT is send and scylla is backed. but any connection
from that node should be ignored.