fix(bootstrapstreamingerror): always remove node after bootstrap abort
Nemesis BootstrapStreamingError could fail because node could not fully
bootstrapped, and during second boot could failed because node with
same ip is in cluster.
To avoid such situation and simplify the nemesis behavior, after each abort
of bootstrap, node will be removed from cluster with removenode command
and all scylla data will be deleted, and the node will be bootstrapped
as new one.

Fixes: #10933